Initial credentials staged (#3)
* lot of new operations

* logical fixes

* minor fix

* login split && readme

+++ models/price.go
@@ -3,13 +3,13 @@ package models
 import "github.com/shopspring/decimal"
 
 type PriceAlert struct {
-	UserId uint
-	GameId uint
+	UserId int64
+	GameId int64
 	Price  decimal.Decimal
 }
 
 type GamePrice struct {
-	GameId       uint
-	StoreId      uint
+	GameId       int64
+	StoreId      int64
 	CurrentPrice decimal.Decimal
 }
